Allow FieldArray access to snr_from_loglr (#1718)
* Allow FieldArray access to snr_from_loglr.

* Update z_args.
